refactor(ai-provider): split cursor pagination helpers into pagination.rs
